Split toolbar from resourceList index
